Updated AVU monitors to 64 bit

+++ accelerators/stratus_hls/audio_ffi_stratus/hw/src/audio_ffi.cpp
@@ -458,12 +458,12 @@ void audio_ffi::compute_kernel()
     int32_t logn_samples;
     int32_t num_samples;
     int32_t do_shift;
-    uint32_t context_quota;
+    uint64_t context_quota;
     uint32_t valid_contexts;
     bool switch_context;
     uint32_t backoff_count;
     int32_t local_input_is_full;
-    uint32_t cycles_elapsed;
+    uint64_t cycles_elapsed;
     {
         HLS_PROTO("compute-config");
 
@@ -480,7 +480,7 @@ void audio_ffi::compute_kernel()
         HLS_FLATTEN_ARRAY(config.do_shift);
         HLS_FLATTEN_ARRAY(config.context_quota);
         
-        context_quota = config.context_quota[current_context_int];
+        context_quota = (uint64_t) config.context_quota[current_context_int];
         valid_contexts = config.valid_contexts;
         logn_samples = config.logn_samples[current_context_int];
         num_samples = 1 << logn_samples;
@@ -509,7 +509,7 @@ void audio_ffi::compute_kernel()
             conf_info_t config = this->conf_info.read();        
             HLS_FLATTEN_ARRAY(config.context_quota);
             
-            context_quota = config.context_quota[current_context_int];
+            context_quota = (uint64_t) config.context_quota[current_context_int];
             valid_contexts = config.valid_contexts;
             wait();
         }
@@ -578,7 +578,7 @@ void audio_ffi::compute_kernel()
             logn_samples = config.logn_samples[current_context_int];
             num_samples = 1 << logn_samples;
             do_shift = config.do_shift[current_context_int];
-            context_quota = config.context_quota[current_context_int];
+            context_quota = (uint64_t) config.context_quota[current_context_int];
             wait();
         }
                 
@@ -1065,9 +1065,9 @@ void audio_ffi::util_monitor()
     }
 
     // Config
-    uint32_t active_cycles[MAX_CONTEXTS];
-    uint32_t start_cycles;
-    uint32_t end_cycles;
+    uint64_t active_cycles[MAX_CONTEXTS];
+    uint64_t start_cycles;
+    uint64_t end_cycles;
     {
         HLS_DEFINE_PROTOCOL("monitor-cfg");
         cfg.wait_for_config(); // config process
